+++ CHANGELOG.md
@@ -13,7 +13,8 @@ Possible log types:
 
 ### Unreleased
 
-- [changed] Update redis to 0.15 and r2d2\_redis to 0.13 (#105)
+- [changed] Update redis to 0.20 and r2d2\_redis to 0.14 (#105, #106)
+- [changed] Update quick-error to 2.0 (#106)
 
 ### v0.5.0 (2020-04-21)